+++ benchmarks/transport_montecarlo_single_packet_loop.py
@@ -7,7 +7,7 @@ from tardis.transport.montecarlo import single_packet_loop
 from numba.np.ufunc.parallel import get_num_threads, get_thread_id
 
 
-class BenchmarkMontecarloMontecarloNumbaVpacket(BenchmarkBase):
+class BenchmarkTransportMontecarloSinglePacketLoop(BenchmarkBase):
     """
     Class to benchmark the single packet loop function.
     """
